# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dokumentasi Sederhana",
        "description": "Dokumentasi API untuk tabel statuses",
        "contact": {
            "email": "raden@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/approvers": {
            "get": {
                "tags": [
                    "Approvers"
                ],
                "summary": "Ambil semua approver",
                "operationId": "9cfc9c40995bc2eea6d10b41bdb8cc54",
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    }
                }
            },
            "post": {
                "tags": [
                    "Approvers"
                ],
                "summary": "Tambah approver",
                "operationId": "fa5a37684f6acea3b6ddbcf152570db2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil dibuat"
                    }
                }
            }
        },
        "/api/v1/approvers/{id}": {
            "get": {
                "tags": [
                    "Approvers"
                ],
                "summary": "Detail approver",
                "operationId": "53a923157456f0f50368dccb6a108e4f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    }
                }
            },
            "put": {
                "tags": [
                    "Approvers"
                ],
                "summary": "Update approver",
                "operationId": "45a417ea17b47b8f95b89937bd174f36",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diupdate"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Approvers"
                ],
                "summary": "Hapus approver",
                "operationId": "57567b8fc49c80f7d0a5742abbdff95a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Berhasil dihapus"
                    }
                }
            }
        },
        "/api/v1/expenses": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "operationId": "5e01e94f14166f3b5b717394a31fa1f9",
                "responses": {
                    "200": {
                        "description": "List of expenses"
                    }
                }
            },
            "post": {
                "tags": [
                    "Expenses"
                ],
                "operationId": "01ede6c11db41a49b7a7bb3e25f213a9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "status_id"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "status_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/v1/expenses/{id}": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "operationId": "52a66329324dd22c9bf790a619e99621",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense detail"
                    }
                }
            },
            "put": {
                "tags": [
                    "Expenses"
                ],
                "operationId": "270309e166d24efe6b9d5333a5ee5b93",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "status_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Expenses"
                ],
                "operationId": "5687d6eb63202c87f6479832a55a3a6c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    }
                }
            }
        },
...
    },
...
}